+++ RouteEngine/Core/Node.h
@@ -8,7 +8,7 @@ private:
     //bu id linkedlist deki TargetNodeId nin ta kendisi
     Point coordinate;       // Karenin X ve Y koordinatı
     bool isObstacle;        // Bu kare bir duvar/engel mi?
-    float distance;         // Dijkstra'nın buraya gelme maliyeti //dıstance o yoldakı weıghtlerın toplamı
+    int distance;         // Dijkstra'nın buraya gelme maliyeti //dıstance o yoldakı weıghtlerın toplamı
     int previousNodeId;     // Rotayı bulduktan sonra geriye doğru çizmek için bir önceki adımın ID'si
 
 public:
@@ -21,7 +21,7 @@ public:
     int GetId();
     Point GetCoordinate();
     bool GetIsObstacle();
-    float GetDistance();
+    int GetDistance();
     int GetPreviousNodeId();
 
     // Değiştirme (Setter) Fonksiyonları
